Created guessTheNumber
Created a game where players guess a number between 1 to 100 where they have 5 tries. Based on each guess the lower bound and higher bound number updates based on if your guess is higher or lower than the number. Then it checks if the difference between the guess and number is within a range of 5, 10, 25 or 50 of the number you are trying to guess. It is based on this the upper and lower bound updates for your next round. Lastly it asks if you want to play again.